requirements.txt
Django==2.1.5
django-crispy-forms==1.7.2
djangorestframework==3.9.1
psycopg2==2.7.7
pytz==2018.9
